Update demo command with the new SimpleCommand, update comment to reflect permission access change
